FAC-935:impl:Move KitLanguage enum to common
