+++ flickit-assessment-common/src/main/java/org/flickit/assessment/common/application/domain/kit/KitLanguage.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.kit.application.domain;
+package org.flickit.assessment.common.application.domain.kit;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Getter;
