+++ flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/application/port/in/assessmentkit/UpdateKitInfoUseCase.java
@@ -6,8 +6,8 @@ import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Value;
 import org.flickit.assessment.common.application.SelfValidating;
+import org.flickit.assessment.common.application.domain.kit.KitLanguage;
 import org.flickit.assessment.common.validation.EnumValue;
-import org.flickit.assessment.kit.application.domain.KitLanguage;
 
 import java.util.List;
 import java.util.UUID;
