+++ flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/application/service/assessmentkit/createbydsl/CreateKitByDslService.java
@@ -2,8 +2,8 @@ package org.flickit.assessment.kit.application.service.assessmentkit.createbydsl
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.flickit.assessment.common.application.domain.kit.KitLanguage;
 import org.flickit.assessment.common.exception.AccessDeniedException;
-import org.flickit.assessment.kit.application.domain.KitLanguage;
 import org.flickit.assessment.kit.application.domain.KitVersionStatus;
 import org.flickit.assessment.kit.application.domain.dsl.AssessmentKitDslModel;
 import org.flickit.assessment.kit.application.port.in.assessmentkit.CreateKitByDslUseCase;
