+++ flickit-assessment-kit/src/test/java/org/flickit/assessment/kit/application/service/assessmentkit/CreateAssessmentKitServiceTest.java
@@ -1,7 +1,7 @@
 package org.flickit.assessment.kit.application.service.assessmentkit;
 
+import org.flickit.assessment.common.application.domain.kit.KitLanguage;
 import org.flickit.assessment.common.exception.AccessDeniedException;
-import org.flickit.assessment.kit.application.domain.KitLanguage;
 import org.flickit.assessment.kit.application.domain.KitVersionStatus;
 import org.flickit.assessment.kit.application.port.in.assessmentkit.CreateAssessmentKitUseCase;
 import org.flickit.assessment.kit.application.port.out.assessmentkit.CreateAssessmentKitPort;
